Improved Topics, Filters, init-script
This commit improves the configuration of topics and filters. The
topics configuration parameter is a Lists that holds all the topics
the importer should subscribe to. The "filters" configuration
paramter is a list of regex-strings that works as a "display" filter.
Only messages that contain the regex-strings will be forwarded to the
unix-domain socket.
This commit also adds a init-script for sys-v-init.